Move Collapse All button to FrontContainer
