+++ client-v2/src/components/FrontsEdit/FrontCollectionsOverview.tsx
@@ -4,16 +4,12 @@ import { styled } from 'constants/theme';
 import { State } from 'types/State';
 import { Dispatch } from 'types/Store';
 import { getFront } from 'selectors/frontsSelectors';
-import { closeCollections } from 'actions/Collections';
 import { FrontConfig } from 'types/FaciaApi';
 import CollectionOverview from './CollectionOverview';
 import { CollectionItemSets } from 'shared/types/Collection';
 import ContainerHeadingPinline from 'shared/components/typography/ContainerHeadingPinline';
 import ContentContainer from 'shared/components/layout/ContentContainer';
-import ButtonCircularWithLabel from 'shared/components/input/ButtonCircularWithLabel';
-import ButtonCircularCaret, {
-  ButtonCircularWithTransition
-} from 'shared/components/input/ButtonCircularCaret';
+import ButtonCircularCaret from 'shared/components/input/ButtonCircularCaret';
 import {
   selectIsFrontOverviewOpen,
   editorOpenOverview,
@@ -27,7 +23,6 @@ interface FrontContainerProps {
 
 type FrontCollectionOverviewProps = FrontContainerProps & {
   front: FrontConfig;
-  closeAllCollections: (collections: string[]) => void;
   overviewIsOpen: boolean;
   toggleOverview: (open: boolean) => void;
 };
@@ -52,27 +47,10 @@ const ContainerBody = styled.div`
   width: 380px;
 `;
 
-const ContainerMeta = styled('div')`
-  margin-top: 10px;
-  width: 100%;
-  display: flex;
-  justify-content: flex-end;
-`;
-
-const CollapseAllButton = styled(ButtonCircularWithLabel)`
-  :hover {
-    ${ButtonCircularWithTransition} {
-      background-color: ${({ theme }) =>
-        theme.shared.button.backgroundColorFocused};
-    }
-  }
-`;
-
 const FrontCollectionsOverview = ({
   id,
   front,
   browsingStage,
-  closeAllCollections,
   overviewIsOpen,
   toggleOverview
 }: FrontCollectionOverviewProps) => (
@@ -90,29 +68,16 @@ const FrontCollectionsOverview = ({
       />
     </ContainerHeadingPinline>
     {overviewIsOpen && (
-      <>
-        <ContainerMeta>
-          <CollapseAllButton
-            onClick={e => {
-              e.preventDefault();
-              closeAllCollections(front.collections);
-            }}
-            label={'Collapse all'}
-          >
-            <ButtonCircularCaret small active preActive={false} />
-          </CollapseAllButton>
-        </ContainerMeta>
-        <ContainerBody>
-          {front.collections.map(collectionId => (
-            <CollectionOverview
-              frontId={id}
-              key={collectionId}
-              collectionId={collectionId}
-              browsingStage={browsingStage}
-            />
-          ))}
-        </ContainerBody>
-      </>
+      <ContainerBody>
+        {front.collections.map(collectionId => (
+          <CollectionOverview
+            frontId={id}
+            key={collectionId}
+            collectionId={collectionId}
+            browsingStage={browsingStage}
+          />
+        ))}
+      </ContainerBody>
     )}
   </Container>
 );
@@ -126,8 +91,6 @@ const mapDispatchToProps = (
   dispatch: Dispatch,
   props: FrontContainerProps
 ) => ({
-  closeAllCollections: (collections: string[]) =>
-    dispatch(closeCollections(collections)),
   toggleOverview: (open: boolean) =>
     dispatch(
       open ? editorOpenOverview(props.id) : editorCloseOverview(props.id)
